agent/requirements.txt
langchain
langchain-openai
langgraph
python-dotenv
pydantic
tavily-python
langchain-community
